// src/typings/yapi/HomeColumnGetListV1GetApi.d.ts
export interface YapiGetV1HomeColumnGetListColumnsDatasData {
  /**
   * 是否外链 1：外链 2：非外链 字典Code：is_link_ind
   */
  isLink: number
  /**
   * 栏目子项List
   */
  childColumns: YapiGetV1HomeColumnGetListChildColumnsListColumnsDatasData[]
  /**
   * 是否父级 1：是 2：否 字典Code：common_is_ind
   */
  isParent: number
  /**
   * 语言类型 字典Code：lan_type_cd
   */
  lanTypeCd: string
  /**
   * web跳转链接
   */
  webUrl: string
  /**
   * 栏目标题名称
   */
  homeColumnName: string
  /**
   * 是否web端展示 1：是 2：否 字典Code：common_is_ind
   */
  isWeb: number
  /**
   * app跳转链接
   */
  appUrl: string
  /**
   * 主键ID
   */
  id: number
  /**
   * 是否app展示 1：是 2：否 字典Code：common_is_ind
   */
  isApp: number
  /**
   * 栏目固定代码：用于卡单独的某个栏目，单独展示 对应数据字典Code：home_column_cd
   */
  homeColumnCd: string
  /**
   * 对应帮助中心_id + types  对应表 mer_help_center.id 此处说明：此处组装为 id-type，比如 返回241-1，若后面的type为1时，则为目录，跳转目录，若type为2时，则为文章，直接跳转文章     id取-前面的值即可
   */
  helpCenterId: string
  /**
   * 排序
   */
  sortCode: string
}
export interface YapiGetV1HomeColumnGetListChildColumnsListColumnsDatasData {
  /**
   * 是否外链 1：外链 2：非外链 字典Code：is_link_ind
   */
  isLink: number
  /**
   * 递归项目子项List 结构跟这一级相同
   */
  childColumns: string[]
  /**
   * 是否父级 1：是 2：否 字典Code：common_is_ind
   */
  isParent: number
  /**
   * 语言类型 字典Code：lan_type_cd
   */
  lanTypeCd: string
  /**
   * 栏目标题名称
   */
  homeColumnName: string
  /**
   * 是否web端展示 1：是 2：否 字典Code：common_is_ind
   */
  isWeb: number
  /**
   * 主键ID
   */
  id: number
  /**
   * 是否app展示 1：是 2：否 字典Code：common_is_ind
   */
  isApp: number
  /**
   * 对应帮助中心_id  对应表 mer_help_center.id
   */
  helpCenterId: number
  /**
   * 栏目固定代码：用于卡单独的某个栏目，单独展示 对应数据字典Code：column_title_name
   */
  homeColumnCd: string
  /**
   * 父级ID 对应上级目录的id字段
   */
  parentId: number
  /**
   * web跳转链接
   */
  webUrl: string
  /**
   * app跳转链接
   */
  appUrl: string
}